Atualização de vitals na gamebars; Alteração de vitals do tipo int para double; Atualização em loop de recuperação de vitals.

+++ External/Database/Migrations/DatabaseContextModelSnapshot.cs
@@ -109,17 +109,17 @@ namespace Database.Migrations
                         .ValueGeneratedOnAdd()
                         .HasColumnType("INTEGER");
 
-                    b.Property<int>("Health")
-                        .HasColumnType("INTEGER");
+                    b.Property<double>("Health")
+                        .HasColumnType("REAL");
 
-                    b.Property<int>("Mana")
-                        .HasColumnType("INTEGER");
+                    b.Property<double>("Mana")
+                        .HasColumnType("REAL");
 
-                    b.Property<int>("MaxHealth")
-                        .HasColumnType("INTEGER");
+                    b.Property<double>("MaxHealth")
+                        .HasColumnType("REAL");
 
-                    b.Property<int>("MaxMana")
-                        .HasColumnType("INTEGER");
+                    b.Property<double>("MaxMana")
+                        .HasColumnType("REAL");
 
                     b.Property<int>("PlayerModelId")
                         .HasColumnType("INTEGER");
